Moved the two functions to index.js
