+++ linkedin-clone/src/actions/index.js
@@ -5,6 +5,7 @@ import db, {
   signInWithPopup,
   createUserWithEmailAndPassword,
 } from "../firebase";
+import { getAuth } from "firebase/auth";
 import { signInWithEmailAndPassword, onAuthStateChanged } from "firebase/auth";
 import store from "../store";
 import { SET_JOB_POSTINGS, SET_USER, SET_USER_JOB_POSTINGS } from "./actionType";
@@ -465,3 +466,45 @@ export function createPosting(jobName, jobField, experience){
     dispatch(addPosting(payload))
   }
 }*/
+
+
+export const filterJobsByPreferences = (jobs, preferences) => {
+  return jobs.filter((job) => {
+    if (!job.jobParameters) {
+      return false;
+    }
+
+    const {
+      experienceLevel,
+      industry,
+      jobType,
+      remoteWorkOption,
+    } = job.jobParameters;
+
+    return (
+      experienceLevel === preferences.experienceLevel &&
+      industry === preferences.industry &&
+      jobType === preferences.jobType &&
+      remoteWorkOption === preferences.remoteWorkOption
+    );
+  });
+};
+
+export   const getUserSearchingPreferences = async () => {
+  try {
+    const currentUser = getAuth().currentUser;
+    const userDocRef = doc(db, 'Users', currentUser.uid);
+    const docSnap = await getDoc(userDocRef);
+    
+    if (docSnap.exists()) {
+      const userData = docSnap.data();
+      return userData.searchingPreferences;
+    } else {
+      console.log('No user found with this userId:', currentUser.uid);
+      return null;
+    }
+  } catch (error) {
+    console.error('Error fetching user preferences:', error);
+    return null;
+  }
+};
